* Added the ability to use filters in if/ifelse expressions, e.g. :
{% if var1|length_is:10 %}

{% ifequal months_list|length %}

* fixed a bug with erlydtl_runtime:are_equal which always
returned false when comparing 2 single character values, e.g. :

  erlydtl_runtime:are_equal("9", "9").
  erlydtl_runtime:are_equal("x", "x").

.. always returned false.  This would have manifested itself in 
the scenario {% ifequal foo "x" %} where foo was a variable with 
the value "x".

* the generated iolist now goes through an additional output 
filter called erlydtl_tuntime:stringify_final.  This will stringify
any atoms making it safe for filters to return atoms and also 
for atoms to be passed in to the render() function.



git-svn-id: http://erlydtl.googlecode.com/svn/trunk@136 a5195066-8e3e-0410-a82a-05b01b1b9875