Issue #11
Rolled back the functionality added in r152 (Issue #9).
The idea of an auto-changing template root folder is going away 
because it is not the "Django way".  The Djnago way is to have
all include paths relative to a base folder (although they do
support many base folders, but I digress).

Trying to support both systems is too complex and introduces unnecessary
ambiguity about the current include path for those doing it the Django way.



git-svn-id: http://erlydtl.googlecode.com/svn/trunk@157 a5195066-8e3e-0410-a82a-05b01b1b9875