Support for i18n at run-time.
Pass a translation dictionary as the second argument to "render"
and it will be used to translate strings inside {% trans %} blocks.
Values in this dictionary will override compile-time translations
achieved with the 'locale' compile-time option.

Also generates a new function translatable_strings/0 which returns
a list of {% trans %} strings than can be converted with a dictionary.