{% blocktrans %} no longer uses identifiers
Previously {% blocktrans %} blocks required an identifier which would be
passed as an atom to the blocktrans_fun function passed in at
compile-time. This behavior was inconsistent with Django. Now identifiers
are no longer used, and the blocktrans_fun function simply receives the
block to be translated as a string.

For convenience, compiled modules now have a translated_blocks/0
function which returns a list of blocks that were translated at
compile-time. A new helper module called blocktrans_extractor is
provided for extracting blocktrans blocks prior to compilation.

The implementation may result in some oddities; the translatable blocks
are actually derived from the abstract syntax tree, so there may
be whitespace or small syntactic differences between the blocks in
the source code and the blocks used at translation-time. However,
the blocktrans_extractor module will return blocks that are exactly
compatible with the translatable blocks. One side-benefit of this
appproach is that non-semantic whitespace will always be ignored;
if two {% blocktrans %} blocks have slight syntactic differences
but are semantically the same, you only need to provide a single
translation to cover the two of them.