Accept any identifier as attribute (Fixes #177)
Now, the . for looking up an attribute of a variable will bypass any other reserved meaning that the
indentifier may normally hold. So, in a way, the attribute lookup opertaion short-circuits the
keyword mechanism. Which makes sense, as it won't make sense to put a keyword after the dot any way.