Fix bug in truncatewords_html on closing tags
`truncatewords_html` works as tested, but wasn't tested against one
potential case of input: if there's a completed closed tag in place.
erlydtl was throwing `function_clause` errors because it expected the
very next characters after `/` to be `>`, when it will normally be the
tag's name (i.e. the characters "strong" in "</strong>").

This adds a test case where truncation doesn't occur, and several tags
are opened and closed, as well as the function clause necessary to parse
the tags correctly.